requirements.txt
streamlit
opencv-python
pillow
numpy
scikit-image
imutils
py-sudoku
